Object lookup map: Replace slist with own implementation
The Zephyr slist implementation did not work on hardware under some
circumstances. The root cause could not be fully identified. It may
have to do with memory alignment or placement of the sys_snode_t at
the end of the struct instead of at the beginning.

This commit replaces the Zephyr linked list with a bespoke
implementation for the ThingSet objects database.